+++ alarmdecodermqtt/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.1.13
+Publish all panel attribute flags on panel topic.  Can be used as json
+attributes when defining the panel.
+
 # 0.1.11
 Add `expander_zones` config setting which allows users to optionally designate
 zones as being on an expander board.  This helps avoid them being erroneously